Begininng hard refactoring
This package, at this point, is beginnign a stage of heavy refactoring
to be much more go friendly.  There are many c-idioms and uneccessary
checks that makes this package really hard to use, and doesnt make use
of useful Goism.

The key part of the library is the netcdf4.go file which imports and
acts as an interface layer between go and NetCDF.  It is currently NOT
thread nor go-routine safe